Explainability/requirements.txt
fastapi 
uvicorn 
python-dotenv 
langchain-openai 
langchain-community 
beautifulsoup4 
lxml
faiss-cpu